+++ app/(tabs)/login.disabled.tsx
@@ -1,0 +1,6 @@
+// login.disabled.tsx
+// Duplicate of the login page kept for reference. This file is intentionally
+// disabled so expo-router won't create a tab route for it. The active login
+// page is at `/app/login.tsx` (root).
+
+export const preserved = true;
