+++ app/(tabs)/register.disabled.tsx
@@ -1,0 +1,6 @@
+// register.disabled.tsx
+// Duplicate of the register page kept for reference. This file is intentionally
+// disabled so expo-router won't create a tab route for it. The active register
+// page is at `/app/register.tsx` (root).
+
+export const preserved = true;
