+++ scripts/archived-auth/login.copy.tsx
@@ -1,0 +1,2 @@
+// Archived copy of app/(tabs)/login.tsx
+export const archived = true;
