+++ scripts/archived-auth/register.copy.tsx
@@ -1,0 +1,2 @@
+// Archived copy of app/(tabs)/register.tsx
+export const archived = true;
