+++ tsconfig.json
@@ -15,4 +15,6 @@
     ".expo/types/**/*.ts",
     "expo-env.d.ts"
   ]
+  ,
+  "exclude": ["backend"]
 }
